feat: Add processFileCRA
