+++ examples/create-react-app/config/jest/fileTransform.js
@@ -1,35 +1,7 @@
-const path = require('path');
-const camelcase = require('camelcase');
+const { processFileCRA } = require('jest-preview');
 
 module.exports = {
   process(src, filename) {
-    // /Users/your-name/your-project/src/assets/image.png => /src/assets/image.png
-    const relativeFilename = JSON.stringify(filename.split(process.cwd())[1]);
-
-    if (filename.match(/\.svg$/)) {
-      // Based on how SVGR generates a component name:
-      // https://github.com/smooth-code/svgr/blob/01b194cf967347d43d4cbe6b434404731b87cf27/packages/core/src/state.js#L6
-      const pascalCaseFilename = camelcase(path.parse(filename).name, {
-        pascalCase: true,
-      });
-      const componentName = `Svg${pascalCaseFilename}`;
-      return `const React = require('react');
-      module.exports = {
-        __esModule: true,
-        default: ${relativeFilename},
-        ReactComponent: React.forwardRef(function ${componentName}(props, ref) {
-          return {
-            $$typeof: Symbol.for('react.element'),
-            type: 'svg',
-            ref: ref,
-            key: null,
-            props: Object.assign({}, props, {
-              children: ${relativeFilename}
-            })
-          };
-        }),
-      };`;
-    }
-    return `module.exports = ${relativeFilename};`;
+    return processFileCRA(src, filename);
   },
 };
