+++ src/transform.ts
@@ -1,4 +1,5 @@
 import path from 'path';
+import camelcase from 'camelcase';
 
 function getRelativeFilename(filename: string): string {
   return filename.split(process.cwd())[1];
@@ -7,6 +8,38 @@ function getRelativeFilename(filename: string): string {
 export function processFile(src: string, filename: string): string {
   const relativeFilename = getRelativeFilename(filename);
   return `module.exports = ${JSON.stringify(relativeFilename)};`;
+}
+
+export function processFileCRA(src: string, filename: string): string {
+  // /Users/your-name/your-project/src/assets/image.png => /src/assets/image.png
+  const relativeFilename = JSON.stringify(getRelativeFilename(filename));
+
+  if (filename.match(/\.svg$/)) {
+    // Based on how SVGR generates a component name:
+    // https://github.com/smooth-code/svgr/blob/01b194cf967347d43d4cbe6b434404731b87cf27/packages/core/src/state.js#L6
+    const pascalCaseFilename = camelcase(path.parse(filename).name, {
+      pascalCase: true,
+    });
+    const componentName = `Svg${pascalCaseFilename}`;
+    return `const React = require('react');
+    module.exports = {
+      __esModule: true,
+      default: ${relativeFilename},
+      ReactComponent: React.forwardRef(function ${componentName}(props, ref) {
+        return {
+          $$typeof: Symbol.for('react.element'),
+          type: 'svg',
+          ref: ref,
+          key: null,
+          props: Object.assign({}, props, {
+            children: ${relativeFilename}
+          })
+        };
+      }),
+    };`;
+  }
+
+  return `module.exports = ${relativeFilename};`;
 }
 
 export function processCss(src: string, filename: string): string {
